1. WebController를 PlayerController, PointController로 분리 2. hbs 파일 일부 수정
