--- bin/main/templates/playerAll.hbs
@@ -1,93 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-<head>
-    <title>Fantasy Baseball</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-
-    <!--부트스트랩 css 추가-->
-    <link rel="stylesheet" href="/css/lib/bootstrap.min.css">
-</head>
-<body>
-    <h1>FBB 서비스 (Fantasy Baseball, ver. 0.3)</h1>
-    <div class="col-md-12">
-        <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#savePlayerModal">선수 등록</button>
-        <br/>
-        <br/>
-        <!-- 목록 출력 영역 -->
-        <table class="table table-horizontal table-bordered">
-            <thead class="thead-strong">
-                <tr>
-                    <th>ID</th>
-                    <th>팀명</th>
-                    <th>선수명</th>
-                    <th>포지션</th>
-                    <th>등번호</th>
-                    <th>연봉</th>
-                </tr>
-            </thead>
-            <tbody id="tbody">
-            {{#each playerAll}}
-                <tr>
-                    <td>{{playerId}}</td>
-                    <td>{{teamName}}</td>
-                    <td>{{playerName}}</td>
-                    <td>{{position}}</td>
-                    <td>{{number}}</td>
-                    <td>{{salary}}</td>
-                </tr>
-            {{/each}}
-            </tbody>
-        </table>
-    </div>
-
-    <!-- Modal 영역 -->
-    <div class="modal fade" id="savePlayerModal" tabindex="-1" role="dialog" aria-labelledby="savePlayerLabel" aria-hidden="true">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h5 class="modal-title" id="savePlayerLabel">선수 등록</h5>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">&times;</span>
-                    </button>
-                </div>
-                <div class="modal-body">
-                    <form>
-                        <div class="form-group">
-                            <label for="teamName"> 팀명 </label>
-                            <input type="text" class="form-control" id="teamName" placeholder="팀명을 입력하세요">
-                        </div>
-                        <div class="form-group">
-                            <label for="playerName"> 선수명 </label>
-                            <input type="text" class="form-control" id="playerName" placeholder="선수명을 입력하세요">
-                        </div>
-                        <div class="form-group">
-                            <label for="position"> 포지션 </label>
-                            <input type="text" class="form-control" id="position" placeholder="포지션을 입력하세요">
-                        </div>
-                        <div class="form-group">
-                            <label for="number"> 등번호 </label>
-                            <input type="text" class="form-control" id="number" placeholder="등번호를 입력하세요">
-                        </div>
-                        <div class="form-group">
-                            <label for="salary"> 연봉 </label>
-                            <input type="text" class="form-control" id="salary" placeholder="연봉을 입력하세요">
-                        </div>
-                    </form>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-dismiss="modal">취소</button>
-                    <button type="button" class="btn btn-primary" id="btn-save">등록</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <!--부트스트랩 js, jquery 추가-->
-    <script src="/js/lib/jquery.min.js"></script>
-    <script src="/js/lib/bootstrap.min.js"></script>
-
-    <script src="/js/app/playerAll.js"></script>
-
-</body>
-</html>
