+++ bin/main/templates/point.hbs
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="/css/lib/bootstrap.min.css">
 </head>
 <body style="background-color: navy; color: white;">
-    <h1><b>&nbsp;&nbsp;FBB 서비스</b> (Fantasy Baseball, ver. 0.6.1)</h1>
+    <h1><b>&nbsp;&nbsp;FBB 점수 목록</b></h1><br/>
     <div class="col-md-12">
         <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#searchPointModal">점수 조회</button>
         <br/>
