+++ src/main/resources/templates/index.hbs
@@ -30,7 +30,7 @@
 	} 
 	function relocate_point()
 	{
-	     location.href = "/findPointAll";
+	     location.href = "/findPointToday";
 	} 
 	</script>
